Add new benchmarks
Specially the "Harder" one for the NormalOrder is very interesting, since it
fails even for the case with size 64. I have some ideas on how to fix this:
Right now we just swap neighboring operators, however we can compute subsequence
of operators that we can swap and swap with another such subsequence on the
right. If we implement a algorithm that take this into account, "Harder" will
become trivial.