Benchmark creation of std::vector<Operator>
The memory usage of std::vector<Operator> scales linearly with increasing the
sizeof(Operator). Therefore, we should this consider when increasing the number of
max_orbitals, or introduce more metadata.

However, in practical exercises such as HubbardKagome, the performance
difference is minimal (around few seconds). And scaling therefore should only
affect tasks that create huge amount of Operators and is memory-bound.